refactoring imagevector to conform to usage pattern in gardener/gardener
